backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langgraph>=0.2.28
tavily-python>=0.5.0
google-genai>=1.0.0
websockets>=13.0
python-dotenv>=1.0.1
